feat: Add sidebar to root layout for navigation
The code changes include adding a new import statement to include the Sidebar component in the root layout. This allows for a sidebar navigation menu to be displayed alongside the main content. The Navbar component and the children components are wrapped inside the Sidebar component.

Recent user commits:
- Revert "20240606 json handler"
- 20240606 json handler
- Take out data folder
- Add gitignore setting for /app/api; Take out api folder from the commit
- Merge branch 'develop' into 20240606_Json-handler
- toast and cookie

Recent repository commits:
- Revert "20240606 json handler"
- 20240606 json handler
- Take out data folder
- Add gitignore setting for /app/api; Take out api folder from the commit
- Merge branch 'develop' into 20240606_Json-handler
- toast and cookie
- chore: Add sharp package for image processing
- chore: Update Next.js configuration and package dependencies
- Update Next.js configuration and package dependencies